ai_service/requirements.txt
langchain==0.1.7
pypdf2>=3.0.1,<3.1.0
pypdf>=4.0.2,<4.1.0
aiokafka>=0.10.0,<0.11.0
redis>=5.0.1,<5.1.0
redis[hiredis]
pdf2image>=1.17.0,<1.18.0
pillow>=10.2.0,<10.3.0
PyMuPDF>=1.23.21,<1.24.0
pytesseract>=0.3.10,<0.4.0
httpx>=0.26.0,<0.27.0